% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{format_percentage}
\alias{format_percentage}
\title{Format Percentage}
\usage{
format_percentage(x, accuracy = 0.01)
}
\arguments{
\item{x}{Numeric vector}

\item{accuracy}{Rounding accuracy (default: 0.01)}
}
\value{
Character vector of formatted values
}
\description{
Helper to format numbers as percentages
}
\keyword{internal}
