% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{generate_simulated_data}
\alias{generate_simulated_data}
\title{Generate Simulated Live Data}
\usage{
generate_simulated_data(symbol, n_days = 30)
}
\description{
Generate Simulated Live Data
}
\keyword{internal}
