% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{make_interactive}
\alias{make_interactive}
\title{Make Plot Interactive}
\usage{
make_interactive(plot, tooltip = "all", dynamicTicks = TRUE, ...)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{tooltip}{Tooltip variables (default: "all")}

\item{dynamicTicks}{Use dynamic ticks (default: TRUE)}

\item{...Additional}{arguments passed to ggplotly}
}
\value{
A plotly object
}
\description{
Convert any ggplot2 object to an interactive plotly visualization.
}
\examples{
\dontrun{
p <- data \%>\% plot_candles()
make_interactive(p)

# Or use pipe
data \%>\%
  plot_equity_curve() \%>\%
  make_interactive()
}
}
