% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_trading}
\alias{scale_fill_trading}
\title{Trading Fill Scale}
\usage{
scale_fill_trading(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_fill_manual}
}
\value{
A ggplot2 fill scale
}
\description{
Fill scale with financial market conventions.
}
